// type.d.ts
export interface Movie {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  first_air_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  budget: number;
  revenue: number;
  status: string;
  genres: [{ id: number; name: string }];
  media_type: string;
  production_companies?: [
    {
      id: number;
      logo_path: string;
      name: string;
      origin_country: string;
    }
  ];
}

export interface Element {
  type:
    | "Trailer"
    | "Featurette"
    | "Clip"
    | "Behind the Scenes"
    | "Teaser"
    | "Bloopers";
}